♻️💥 Refactor everything & improve user experience (#2)
* :bug: Force adventure mode

* :wrench: IDE stuff

* :construction: Add version property to times

* :page_facing_up: Update LICENSE

* :construction: Reformatting & improvements

* :fire: Database entity for locations

* :recycle: Store parkours in a map associated by its id

* :bulb: Remove unnecessary todo

* :recycle: Refactor menu item

* :recycle: Refactor commands

* :recycle: Refactor listeners

* :recycle: Refactor utils

* :recycle: Refactor item related stuff

* :recycle: Convert playerData to HashMap

* :recycle: Revert version dependent stuff

* :bug: Fix minor bugs

* :boom: Store ticks as long

* :recycle: Simplify mine-cart interaction

* 🔊 Add new messages

* :bug: Fix bug when adding a new parkour

+++ src/main/kotlin/net/rebux/jumpandrun/database/models/Times.kt
@@ -4,8 +4,9 @@ import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.javatime.datetime
 
 object Times : IntIdTable() {
-    val uuid = uuid("uuid")
-    val time = integer("time")
-    val date = datetime("date")
+
+    val uuid    = uuid("uuid")
+    val time    = long("time")
+    val date    = datetime("date")
     val parkour = reference("parkour", Parkours)
 }
